Move styled components to styles folder
