+++ packages/message-bar/__tests__/message-manager.base.js
@@ -1,10 +1,11 @@
 import React from "react";
-import { View, Text } from "react-native";
+import { TcView, TcText } from "@times-components/utils";
 import TestRenderer from "react-test-renderer";
 import { scales } from "@times-components/ts-styleguide";
 import { delay } from "@times-components/test-utils";
 import MessageManager from "../src/message-manager";
-import MessageBar, { CloseButton } from "../src/message-bar";
+import MessageBar from "../src/message-bar";
+import { CloseButton } from '../src/styles';
 import Context from "../src/message-context";
 
 const Button = () => <button type="button" />;
@@ -23,9 +24,9 @@ export default animate => [
     test: async () => {
       const testInstance = TestRenderer.create(
         <MessageManager delay={1} scale={scales.medium}>
-          <View>
-            <Text>test child content</Text>
-          </View>
+          <TcView>
+            <TcText>test child content</TcText>
+          </TcView>
         </MessageManager>
       );
